Extract canvas resize
